set refresh_token on instance retrieval.
Apparently, toBeDefined doesn't check for null
And we had a test asserting that the credential.refresh_token existed, which is good, but it wasn't getting passed along on new instance.
The refresh token test was working because it was being done on the freshly authenticated manager.api class (vs. a fresh retrieval).