Update the integration-factory to take advantage of/utilize the new methods for the integration-base.
Add the entityReference used by the integration-factory and any new itnegrations added, which allows for name-based module reference. This allows for the developer importing the same api module multiple times but with different intended purpose. Example: slack-admin and slack-user. There may be other use cases as well.

This likely should be paired with a few other integration-related updates.